Quorum_backend::cli

Module server

Functions

audit

Displays server audit logs

login

Handles user login

logout

Handles user logout

logs

Displays server logs

make_admin

Handles making a user an admin

prompt 🔒

Prompts the user for input with a given label and returns the trimmed input as a String.

prompt_password 🔒

Prompts the user for a password with a given label and returns the input as a String.

reload

Handles reloading the server configuration

shutdown

Handles server shutdown

signup

Handles user signup

status

Displays server status