Quorum_backend::db::queries

Module auth

Authentication related database queries

This file contains functions for Creating an account (signup), login, token management and deleteing a user account.

Functions

delete_user_by_id

Deletes a user from the database by their ID

get_user_by_username_or_email

Retrieves a user from the database by their username or email

revoke_refresh_token

Revokes a refresh token, preventing it from being used for future authentication

signup_user

Creates a new user account in the database

store_refresh_token

Stores a refresh token in the database for a specific user

validate_refresh_token

Validates a refresh token for a specific user